api/db_handler/requirements.txt
uuid
flask
getmac
psycopg2-binary
python-dotenv
requests